refactor(core): add clipboard Cargo feature, enhancing binary size
