+++ .changes/clipboard-feature.md
@@ -1,0 +1,7 @@
+---
+"tauri": patch
+"tauri-runtime": minor
+"tauri-runtime-wry": minor
+---
+
+**Breaking change::* Added the `clipboard` Cargo feature.
