+++ core/tauri-runtime-wry/Cargo.toml
@@ -42,3 +42,4 @@ objc-exception = [ "wry/objc-exception" ]
 gtk-tray = [ "wry/gtk-tray" ]
 ayatana-tray = [ "wry/ayatana-tray" ]
 global-shortcut = [ "tauri-runtime/global-shortcut" ]
+clipboard = [ "tauri-runtime/clipboard" ]
