+++ core/tauri-runtime/Cargo.toml
@@ -47,3 +47,4 @@ devtools = [ ]
 system-tray = [ ]
 macos-private-api = [ ]
 global-shortcut = [ ]
+clipboard = [ ]
