+++ core/tauri-runtime/src/lib.rs
@@ -311,6 +311,7 @@ pub trait GlobalShortcutManager: Debug + Clone + Send + Sync {
 }
 
 /// Clipboard manager.
+#[cfg(feature = "clipboard")]
 pub trait ClipboardManager: Debug + Clone + Send + Sync {
   /// Writes the text into the clipboard as plain text.
   fn write_text<T: Into<String>>(&mut self, text: T) -> Result<()>;
@@ -332,6 +333,7 @@ pub trait Runtime<T: UserEvent>: Debug + Sized + 'static {
   #[cfg(feature = "global-shortcut")]
   type GlobalShortcutManager: GlobalShortcutManager;
   /// The clipboard manager type.
+  #[cfg(feature = "clipboard")]
   type ClipboardManager: ClipboardManager;
   /// The tray handler type.
   #[cfg(feature = "system-tray")]
@@ -358,6 +360,7 @@ pub trait Runtime<T: UserEvent>: Debug + Sized + 'static {
   fn global_shortcut_manager(&self) -> Self::GlobalShortcutManager;
 
   /// Gets the clipboard manager.
+  #[cfg(feature = "clipboard")]
   fn clipboard_manager(&self) -> Self::ClipboardManager;
 
   /// Create a new webview window.
