+++ core/tauri/Cargo.toml
@@ -142,6 +142,7 @@ fs-extract-api = [ "zip" ]
 reqwest-client = [ "reqwest", "bytes" ]
 process-command-api = [ "shared_child", "os_pipe" ]
 global-shortcut = [ "tauri-runtime/global-shortcut", "tauri-runtime-wry/global-shortcut" ]
+clipboard = [ "tauri-runtime/clipboard", "tauri-runtime-wry/clipboard" ]
 dialog = [ "rfd" ]
 notification = [ "notify-rust" ]
 cli = [ "clap" ]
@@ -168,8 +169,8 @@ api-all = [
   "window-all"
 ]
 clipboard-all = [ "clipboard-write-text", "clipboard-read-text" ]
-clipboard-read-text = [ ]
-clipboard-write-text = [ ]
+clipboard-read-text = [ "clipboard" ]
+clipboard-write-text = [ "clipboard" ]
 dialog-all = [ "dialog-open", "dialog-save", "dialog-message", "dialog-ask" ]
 dialog-ask = [ "dialog" ]
 dialog-confirm = [ "dialog" ]
